location geometry added

+++ api-amanda/DTOs/TrackInfoDTO.cs
@@ -1,7 +1,7 @@
 ﻿namespace api_amanda.DTOs {
     public class TrackInfoDTO:CsvRecordDTO {
-        public decimal btsLat { get; set; }
-        public decimal btsLon { get; set; }
+        public double btsLat { get; set; }
+        public double btsLon { get; set; }
 
     }
 }
